chore: self review refactor
